@@ -12,7 +12,7 @@ module RSpec
         case method_ref.visibility
         when :public    then true
         when :private   then include_private
-        when :protected then include_private || RUBY_VERSION.to_f < 2.0
+        when :protected then include_private
         else !method_ref.unimplemented?
         end
       end
