@@ -93,7 +93,6 @@ module RSpec
       end
 
       it 'allows a class and a subclass to both be stubbed' do
-        pending "Does not work on 1.8.7 due to singleton method restrictions" if RUBY_VERSION == "1.8.7" && RSpec::Support::Ruby.mri?
         the_klass = Class.new
         the_subklass = Class.new(the_klass)
 
@@ -181,20 +180,10 @@ module RSpec
         file_2.close
       end
 
-      def expect_output_warning_on_ruby_lt_2
-        if RUBY_VERSION >= '2.0'
-          yield
-        else
-          expect { yield }.to output(a_string_including(
-            "RSpec could not fully restore", file_1.inspect, "write"
-          )).to_stderr
-        end
-      end
-
-      it "allows `write` to be stubbed and reset", :unless => Support::Ruby.jruby? do
+      it "allows `write` to be stubbed and reset" do
         allow(file_1).to receive(:write)
         file_1.reopen(file_2)
-        expect_output_warning_on_ruby_lt_2 { reset file_1 }
+        reset file_1
 
         expect {
           # Writing to a file that's been reopened to a 2nd file writes to both files.
@@ -204,7 +193,7 @@ module RSpec
       end
     end
 
-    RSpec.describe "Using a partial mock on a proxy object", :if => defined?(::BasicObject) do
+    RSpec.describe "Using a partial mock on a proxy object" do
       let(:proxy_class) do
         Class.new(::BasicObject) do
           def initialize(target)
