@@ -83,60 +83,58 @@ module RSpec
         }.to raise_error(ArgumentError, "Wrong number of arguments. Expected 0, got 1.")
       end
 
-      if required_kw_args_supported?
-        it 'allows keyword arguments' do
-          o = instance_double('LoadedClass', :kw_args_method => true)
-          expect(o.kw_args_method(1, :required_arg => 'something')).to eq(true)
+      it 'allows keyword arguments' do
+        o = instance_double('LoadedClass', :kw_args_method => true)
+        expect(o.kw_args_method(1, :required_arg => 'something')).to eq(true)
+      end
+
+      context 'for a method that only accepts keyword args' do
+        it 'allows hash matchers like `hash_including` to be used in place of the keywords arg hash' do
+          o = instance_double('LoadedClass')
+          expect(o).to receive(:kw_args_method).
+            with(1, hash_including(:required_arg => 1))
+          o.kw_args_method(1, :required_arg => 1)
         end
 
-        context 'for a method that only accepts keyword args' do
-          it 'allows hash matchers like `hash_including` to be used in place of the keywords arg hash' do
-            o = instance_double('LoadedClass')
+        it 'allows anything matcher to be used in place of the keywords arg hash' do
+          o = instance_double('LoadedClass')
+          expect(o).to receive(:kw_args_method).with(1, anything)
+          o.kw_args_method(1, :required_arg => 1)
+        end
+
+        it 'still checks positional arguments when matchers used for keyword args' do
+          o = instance_double('LoadedClass')
+          prevents(/Expected 1, got 3/) {
             expect(o).to receive(:kw_args_method).
-              with(1, hash_including(:required_arg => 1))
-            o.kw_args_method(1, :required_arg => 1)
-          end
-
-          it 'allows anything matcher to be used in place of the keywords arg hash' do
-            o = instance_double('LoadedClass')
-            expect(o).to receive(:kw_args_method).with(1, anything)
-            o.kw_args_method(1, :required_arg => 1)
-          end
-
-          it 'still checks positional arguments when matchers used for keyword args' do
-            o = instance_double('LoadedClass')
-            prevents(/Expected 1, got 3/) {
-              expect(o).to receive(:kw_args_method).
-                with(1, 2, 3, hash_including(:required_arg => 1))
-            }
-            reset o
-          end
-
-          it 'does not allow matchers to be used in an actual method call' do
-            o = instance_double('LoadedClass')
-            matcher = hash_including(:required_arg => 1)
-            allow(o).to receive(:kw_args_method).with(1, matcher)
-            expect {
-              o.kw_args_method(1, matcher)
-            }.to raise_error(ArgumentError)
-          end
+              with(1, 2, 3, hash_including(:required_arg => 1))
+          }
+          reset o
         end
 
-        context 'for a method that accepts a mix of optional keyword and positional args' do
-          it 'allows hash matchers like `hash_including` to be used in place of the keywords arg hash' do
-            o = instance_double('LoadedClass')
-            expect(o).to receive(:mixed_args_method).with(1, 2, hash_including(:optional_arg_1 => 1))
-            o.mixed_args_method(1, 2, :optional_arg_1 => 1)
-          end
-        end
-
-        it 'checks that stubbed methods with required keyword args are ' \
-           'invoked with the required arguments' do
-          o = instance_double('LoadedClass', :kw_args_method => true)
+        it 'does not allow matchers to be used in an actual method call' do
+          o = instance_double('LoadedClass')
+          matcher = hash_including(:required_arg => 1)
+          allow(o).to receive(:kw_args_method).with(1, matcher)
           expect {
-            o.kw_args_method(:optional_arg => 'something')
+            o.kw_args_method(1, matcher)
           }.to raise_error(ArgumentError)
         end
+      end
+
+      context 'for a method that accepts a mix of optional keyword and positional args' do
+        it 'allows hash matchers like `hash_including` to be used in place of the keywords arg hash' do
+          o = instance_double('LoadedClass')
+          expect(o).to receive(:mixed_args_method).with(1, 2, hash_including(:optional_arg_1 => 1))
+          o.mixed_args_method(1, 2, :optional_arg_1 => 1)
+        end
+      end
+
+      it 'checks that stubbed methods with required keyword args are ' \
+         'invoked with the required arguments' do
+        o = instance_double('LoadedClass', :kw_args_method => true)
+        expect {
+          o.kw_args_method(:optional_arg => 'something')
+        }.to raise_error(ArgumentError)
       end
 
       it 'validates `with` args against the method signature when stubbing a method' do
@@ -213,18 +211,10 @@ module RSpec
           expect(obj.respond_to?(:defined_private_method, false)).to be false
         end
 
-        if RUBY_VERSION.to_f < 2.0 # respond_to?(:protected_method) changed behavior in Ruby 2.0.
-          it 'reports that it responds to protected methods' do
-            expect(obj.respond_to?(:defined_protected_method)).to be true
-            expect(obj.respond_to?(:defined_protected_method, true)).to be true
-            expect(obj.respond_to?(:defined_protected_method, false)).to be true
-          end
-        else
-          it 'reports that it responds to protected methods when the appropriate arg is passed' do
-            expect(obj.respond_to?(:defined_protected_method)).to be false
-            expect(obj.respond_to?(:defined_protected_method, true)).to be true
-            expect(obj.respond_to?(:defined_protected_method, false)).to be false
-          end
+        it 'reports that it responds to protected methods when the appropriate arg is passed' do
+          expect(obj.respond_to?(:defined_protected_method)).to be false
+          expect(obj.respond_to?(:defined_protected_method, true)).to be true
+          expect(obj.respond_to?(:defined_protected_method, false)).to be false
         end
       end
     end
